attempt add moving fixtures around
